services/ai-engine/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
pydantic>=2.7.4
python-dotenv>=1.0.1
pymongo>=4.7.3
